Sort persistence even when toggling different category filter (read description)
However, I feel that current code might not be the most optimised as the filter logic first checks the sorting status, filters the catalog then re-sorts the filtered catalog again.

Also, a lot of code repeats especially for sorting (it is repeated in filter code) as it is not written in Object-oriented style

Lastly, filter and sort is still rather loosely coupled with search function. Ideally, after initial search (no filters, default random sorting), we want to be able to filter and sort the search result too. Will work on that soon